+++ lib/Circle.h
@@ -3,6 +3,7 @@
 
 #include "Point.h"
 #include "Polygon.h"
+#include <GL/gl.h>
 
 class Circle : public Polygon{
 public:
@@ -10,7 +11,7 @@ public:
 	double radius;
 	Circle();
 	Circle(double radius);
-	Circle (double radius, double color[3]);
+	Circle (double radius, GLfloat color[3]);
 	Circle(Circle* circle);
 	void draw();
 	bool in(double x, double y, double r);
